gps-server-master/requirements.txt
aiohttp==3.3.2
aiokafka==0.4.0
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
geographiclib==1.49
geopy==1.16.0
idna==2.7
kafka-python==1.3.5
multidict==4.3.1
python-consul==1.1.0
requests==2.19.1
six==1.11.0
ujson==1.35
urllib3==1.23
uvloop==0.9.1
yarl==1.2.6
